DescriptorHeap min ReShade version implemented. Updated to ReShade 6.0.0 Added a minimal required ReShade version for descriptor heaps to be reset by out addon. Updated to ReShade 6.0.0

+++ gamebridge_reshade/src/igraphicsapi.h
@@ -28,9 +28,9 @@ struct Destroy_Resource_Data
 
 class IGraphicsApi {
 public:
-    size_t reshade_version_nr_major = 0;
-    size_t reshade_version_nr_minor = 0;
-    size_t reshade_version_nr_patch = 0;
+    int32_t reshade_version_nr_major = 0;
+    int32_t reshade_version_nr_minor = 0;
+    int32_t reshade_version_nr_patch = 0;
 
     virtual void draw_debug_overlay(reshade::api::effect_runtime* runtime) = 0;
     virtual void draw_sr_settings_overlay(reshade::api::effect_runtime* runtime) = 0;
